creatDoc
