+++ front/src/component/DoctorList/styles.js
@@ -1,0 +1,14 @@
+import { makeStyles } from "@material-ui/core";
+
+const useStyles = makeStyles((theme) => ({
+  
+  mainContent: {
+  color: 'green',
+  },
+  cardMedia: {
+    paddingTop: "100%",
+  }
+
+}));
+
+export default useStyles;
